updated feature tests names to be prefixed with feature
