+++ tests/Feature/Project/ProjectCreateTest.php
@@ -11,7 +11,7 @@ use App\User;
 use Faker\Factory as Faker;
 use Auth;
 
-class CreateTest extends TestCase
+class ProjectCreateTest extends TestCase
 {
     use DatabaseTransactions;
     protected $team;
