+++ tests/Feature/Section/SectionDeleteTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feaure\Project;
+namespace Tests\Feaure\Section;
 
 use App\UserTeam;
 use Tests\TestCase;
@@ -9,12 +9,14 @@ use Laravel\Passport\Passport;
 use App\Team;
 use App\User;
 use App\Project;
+use App\Section;
 use Auth;
 
-class DeleteTest extends TestCase
+class SectionDeleteTest extends TestCase
 {
     use DatabaseTransactions;
     protected $team;
+    protected $project;
 
     protected function setUp(){
         parent::setUp();
@@ -29,57 +31,61 @@ class DeleteTest extends TestCase
             'user_id' => Auth::user()->id,
             'team_id' => $this->team->id,
         ]);
+        /** create new project */
+        $this->project = factory(Project::class)->create([
+            'team_id' => $this->team->id
+        ]);
     }
 
     /**
-     * Tests ProjectController::destroy
+     * Tests Route section.delete
      *
      * @test
      */
-    public function can_delete_project()
+    public function can_delete_project_section()
     {
         /** Arrange */
-        /** create new project */
-        $project = factory(Project::class)->create([
-            'team_id' => $this->team->id
+        /** create new project section*/
+        $section = factory(Section::class)->create([
+            'project_id' => $this->project->id
         ]);
         /** Act */
-        $response = $this->json('DELETE', "/api/team/".$this->team->id."/project/".$project->id);
+        $response = $this->json('DELETE', "/api/team/".$this->team->id."/project/".$this->project->id."/section/".$section->id);
         /** Assert response is correct */
         $response->assertStatus(200)
             ->assertJson([
                 'success' => true,
-                'message' => 'Project '.$project->name.' has been successfully deleted'
+                'message' => 'Section '.$section->name.' has been successfully deleted'
             ]);
         /** assert project is soft deleted */
-        $this->assertSoftDeleted('projects', [
-            'id' => $project->id
+        $this->assertSoftDeleted('sections', [
+            'id' => $section->id
         ]);
     }
 
     /**
-     * Tests ProjectController::destroy
+     * Tests Route section.delete
      *
      * @test
      */
-    public function cannot_delete_project_in_team_user_is_not_a_member_of()
+    public function cannot_delete_project_section_in_team_user_is_not_a_member_of()
     {
         /** Arrange */
-        /** create new project */
-        $project = factory(Project::class)->create([
-            'team_id' => $this->team->id
+        /** create new project section*/
+        $section = factory(Section::class)->create([
+            'project_id' => $this->project->id
         ]);
         /** create new user and don't add to team*/
         Passport::actingAs(
             factory(User::class)->create()
         );
         /** Act */
-        $response = $this->json('DELETE', "/api/team/".$this->team->id."/project/".$project->id);
+        $response = $this->json('DELETE', "/api/team/".$this->team->id."/project/".$this->project->id."/section/".$section->id);
         /** Assert response is correct */
         $response->assertStatus(403);
-        /** assert project is not deleted */
-        $this->assertDatabaseHas('projects', [
-            'id' => $project->id
+        /** assert section is not deleted */
+        $this->assertDatabaseHas('sections', [
+            'id' => $section->id
         ]);
     }
